refactor(tests): move pipeline/notify tests to tests/core/ subdirs
Move pipeline batch and encode tests to tests/core/pipeline/.
Move notify helper and channel tests to tests/core/notify/.
Make PipelineQuery, encode_pipeline, validate_channel_name,
quote_identifier, quote_literal pub for external test access.
Add Default impl for ConnectionMeta.
