+++ tests/core/mod.rs
@@ -4,7 +4,8 @@ mod cancel;
 mod config;
 mod copy_binary;
 mod copy_text;
-mod notify_channel;
+mod notify;
+mod pipeline;
 mod pool;
 mod protocol_backend;
 mod protocol_codec;
