+++ tests/core/notify/mod.rs
@@ -1,0 +1,2 @@
+mod channel;
+mod helpers;
