+++ tests/core/pipeline/mod.rs
@@ -1,0 +1,2 @@
+mod batch;
+mod encode;
